services/asset-service/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
prometheus-fastapi-instrumentator==7.0.0
